+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.4.0...main)
   - Bugfix: Event IRacingCarInfo now sets LocalUser correctly
   - Event: Added CurrentDriverLicense to IRacingCarInfo
-- - Event: Added Category to IRacingCurrentSession
+  - Event: Added Category to IRacingCurrentSession
+  - Event: Rename LuaManagerCommandDeduplicateEvents to LuaCommandDeduplicateEvents
 
 ## [0.4.0](https://github.com/dennis/slipstream/releases/tag/v0.3.0) (2020-01-10)
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.3.0...v0.4.0)